feat(passthrough): add FreeBSD support
Implements the FreeBSD port of server/passthrough. Same anchored-fd
security model as Linux: every Node holds an O_PATH fd, every
name-based operation goes through (parentFd, name) anchored *at calls.

Notable changes:

* O_PATH constant -- golang.org/x/sys exports unix.O_PATH only for
  freebsd/riscv64 as of v0.42.0. Define a local oPath constant in
  three per-build files (linux, freebsd && riscv64, freebsd &&
  !riscv64) sourcing the flag from unix.O_PATH where available and
  from the literal value (0x400000) on the architectures the upstream
  generator missed. FreeBSD 14+ has the same value across all
  architectures (per <sys/fcntl.h>). All passthrough source now uses
  the bare oPath identifier.

* /proc/self/fd elimination -- replace the Linkat AT_SYMLINK_FOLLOW
  trick in Link, the Lchown reopen in Setattr, and the UtimesNanoAt
  reopen in Setattr with parent-anchored *at calls (Linkat(srcParentFd,
  ...), Fchownat(parentFd, ..., AT_SYMLINK_NOFOLLOW), UtimesNanoAt(
  parentFd, ..., AT_SYMLINK_NOFOLLOW)). procfs is not mounted on
  FreeBSD by default and AT_EMPTY_PATH is unavailable. The
  (parentFd, name) pair preserves TOCTOU resistance. These rewrites
  also benefit Linux -- one less /proc dependency. Root nodes (no
  parentFd) fall back to fd-direct ops or stored host path.

* Setattr Timespec construction goes through unix.NsecToTimespec so
  the field types resolve correctly across 32- and 64-bit Timespec
  layouts (FreeBSD/386 has int32 Sec, FreeBSD/arm has mixed
  int64+int32, etc.).

* FreeBSD dirent parser -- Getdirentries returns a 24-byte header
  (vs Linux's 19) with explicit Namlen field. Hand-rolled parser per
  the ztypes_freebsd struct layout.

* FreeBSD xattr -- the *Fd variants (ExtattrGetFd/SetFd/DeleteFd/
  ListFd) with namespace-prefix decomposition (user./system. ->
  namespace constant + bare name). Length-prefix-encoded list parser.
  SYSTEM namespace EPERM is swallowed for unprivileged servers.

* Statfs_t differences -- Bavail signed -> uint64, Ffree signed ->
  uint64, Namelen renamed to Namemax, Type already uint32.

* Stat_t differences -- Mode uint16 -> cast to uint32; Atim/Mtim/Ctim
  field names match Linux.

* Mknodat dev type drift -- FreeBSD wants uint64; Linux wants int.

handle.go and lock.go: build tag broadened to linux || freebsd
(bodies are portable). passthrough_test.go and kernel_test.go remain
linux-only; FreeBSD test execution is deferred to follow-up (no
FreeBSD CI runner in this repo).

Cross-platform compile verified for darwin/{amd64,arm64} and
windows/{amd64,arm64} (unchanged) and for freebsd/{amd64,arm64,386,
arm,riscv64} (new).

+++ server/passthrough/stat_freebsd.go
@@ -1,4 +1,4 @@
-//go:build linux
+//go:build freebsd
 
 package passthrough
 
@@ -12,12 +12,15 @@ import (
 
 // statToAttr converts a unix.Stat_t to proto.Attr with all fields mapped.
 // UID/GID are transformed through mapper.FromHost for protocol-level reporting.
+//
+// FreeBSD's Stat_t.Mode is uint16 (vs Linux's uint32) and Blksize is int32
+// (vs Linux's int64); explicit casts widen to the proto.Attr field types.
 func statToAttr(st *unix.Stat_t, mapper UIDMapper) proto.Attr {
 	uid, gid := mapper.FromHost(st.Uid, st.Gid)
 	return proto.Attr{
 		Valid:     proto.AttrAll,
 		QID:       statToQID(st),
-		Mode:      st.Mode,
+		Mode:      uint32(st.Mode),
 		UID:       uid,
 		GID:       gid,
 		NLink:     st.Nlink,
@@ -39,7 +42,7 @@ func statToAttr(st *unix.Stat_t, mapper UIDMapper) proto.Attr {
 // number.
 func statToQID(st *unix.Stat_t) proto.QID {
 	var t proto.QIDType
-	switch st.Mode & unix.S_IFMT {
+	switch uint32(st.Mode) & unix.S_IFMT {
 	case unix.S_IFDIR:
 		t = proto.QTDIR
 	case unix.S_IFLNK:
@@ -55,10 +58,8 @@ func statToQID(st *unix.Stat_t) proto.QID {
 }
 
 // toProtoErr converts an OS error to a proto.Errno via proto.ErrnoFromUnix
-// (identity on Linux; FreeBSD-to-Linux wire mapping on FreeBSD). unix.Errno
-// is a type alias for syscall.Errno on Linux, so errors.AsType[unix.Errno]
-// also matches the syscall.Errno wrapped by os.PathError / os.OpenFile.
-// Unknown errors map to EIO. Returns nil for nil input.
+// (which translates FreeBSD errno values to their Linux wire equivalents).
+// Returns nil for nil input.
 func toProtoErr(err error) error {
 	if err == nil {
 		return nil
